+++ substrate/frame/support/test/tests/runtime_metadata.rs
@@ -23,7 +23,7 @@ use sp_metadata_ir::{
 use sp_runtime::traits::Block as BlockT;
 
 pub type BlockNumber = u64;
-pub type Header = sp_runtime::generic::Header<u32, sp_runtime::traits::BlakeTwo256>;
+pub type Header = sp_runtime::generic::Header<u64, sp_runtime::traits::BlakeTwo256>;
 pub type Block = sp_runtime::generic::Block<Header, UncheckedExtrinsic>;
 pub type UncheckedExtrinsic = sp_runtime::generic::UncheckedExtrinsic<u32, RuntimeCall, (), ()>;
